requirements.txt
langchain==0.3.0
langchain-anthropic==0.2.0
langchain-openai==0.2.0
langchain-community==0.3.0
chromadb==0.5.0
pypdf==4.3.1
python-dotenv==1.0.1
rich==13.7.1
typing-extensions==4.12.2